@@ -23,8 +23,8 @@
   "license": "GPL-2.0+",
   "require": {
     "ext-Phar": "*",
-    "php": ">=5.5.0",
-    "typo3/cms-core": "^7.6 || ^8.0.0"
+    "php": ">=7.0.0 <7.2.0",
+    "typo3/cms-core": "^8.0.0"
   },
   "autoload": {
     "psr-4": {
