@@ -1,44 +1,24 @@
 <?php
-
-/***************************************************************
- * Extension Manager/Repository config file for ext "t3adminer".
- *
- * Auto generated 26-05-2013 18:24
- *
- * Manual updates:
- * Only the data in the array - everything else is removed by next
- * writing. "version" and "dependencies" must not be touched!
- ***************************************************************/
-
 $EM_CONF[$_EXTKEY] = array(
 	'title' => 'Adminer',
 	'description' => 'Database administration tool \'Adminer\'',
 	'category' => 'module',
 	'author' => 'Jigal van Hemert',
 	'author_email' => 'jigal.van.hemert@typo3.org',
-	'shy' => '',
-	'dependencies' => '',
-	'conflicts' => '',
-	'priority' => '',
+	'author_company' => '',
 	'module' => 'mod1',
 	'state' => 'stable',
-	'internal' => '',
 	'uploadfolder' => 0,
 	'createDirs' => '',
-	'modify_tables' => '',
 	'clearCacheOnLoad' => 0,
-	'lockType' => '',
-	'author_company' => '',
-	'version' => '7.0.10',
+	'version' => '8.0.0',
 	'constraints' => array(
 		'depends' => array(
-			'typo3' => '7.6.0-8.2.99',
+			'typo3' => '8.7.0-8.7.99',
 		),
 		'conflicts' => array(
 		),
 		'suggests' => array(
 		),
-	),
-	'suggests' => array(
 	),
 );
