@@ -1,7 +1,7 @@
 <?php
 defined('TYPO3_MODE') or die();
 
-if (TYPO3_MODE == 'BE') {
+if (TYPO3_MODE === 'BE') {
 	\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addModule(
 		'tools',
 		'txt3adminerM1',
@@ -17,6 +17,7 @@ if (TYPO3_MODE == 'BE') {
 				),
 				'll_ref' => 'LLL:EXT:t3adminer/mod1/locallang_mod.xml',
 			),
+            'icon' => 'EXT:t3adminer/Resources/Public/Icons/module-adminer.svg',
 			'ADM_subdir' => 'res/',
 			'ADM_script' => 't3adminer.php',
 		)
