@@ -1,8 +1,9 @@
 <?php
 // Make instance:
+use jigal\t3adminer\Controller\AdminerController;
+
 /** @var $SOBE jigal\t3adminer\Controller\AdminerController */
-$SOBE = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('jigal\\t3adminer\\Controller\\AdminerController');
+$SOBE = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(AdminerController::class);
 $SOBE->init();
 
 $SOBE->main();
-$SOBE->printContent();
