@@ -2,7 +2,7 @@
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2011-2014 Jigal van Hemert <jigal.van.hemert@typo3.org>
+ *  (c) 2011-2017 Jigal van Hemert <jigal.van.hemert@typo3.org>
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is
